Struct organelle::Axon

pub struct Axon<T: Soma + 'static> { /* fields omitted */ }

wrap a soma with a set of requirements that will be validated upon startup

Methods

impl<T: Soma + 'static> Axon<T>

pub fn new(
soma: T,
dendrites: Vec<Constraint<T::Synapse>>,
terminals: Vec<Constraint<T::Synapse>>
) -> Self

wrap a soma with constraints specified by dendrite and terminal constraints