Struct organelle::Organelle

pub struct Organelle<T: Soma> where
    T: Soma,  { /* fields omitted */ }

a soma designed to facilitate connections between other somas

where somas are the single cells of functionality, organelles are the organisms capable of more complex tasks. however, organelles are still essentially somas, so they can used in larger organelles as long as they comply with their standards.

Methods

impl<T: Soma + 'static> Organelle<T>

pub fn new(main: T, handle: Handle) -> Self

create a new organelle

pub fn nucleus(&self) -> Uuid

get the main soma's uuid

pub fn add_soma<U: Soma + 'static>(&mut self, soma: U) -> Uuid where
U::Synapse: From<T::Synapse> + Into<T::Synapse>,
<U::Synapse as Synapse>::Dendrite: From<<T::Synapse as Synapse>::Dendrite> + Into<<T::Synapse as Synapse>::Dendrite>,
<U::Synapse as Synapse>::Terminal: From<<T::Synapse as Synapse>::Terminal> + Into<<T::Synapse as Synapse>::Terminal>,

add a soma to the organelle

pub fn connect(
&self,
dendrite: Uuid,
terminal: Uuid,
synapse: T::Synapse
) -> Result<()>

connect two somas together using the specified synapse

pub fn add_dendrite(
&self,
dendrite: (Uuid, <T::Synapse as Synapse>::Dendrite),
terminal: Uuid,
synapse: T::Synapse
) -> Result<()>

send a dendrite to the specified soma

pub fn add_terminal(
&self,
terminal: (Uuid, <T::Synapse as Synapse>::Terminal),
dendrite: Uuid,
synapse: T::Synapse
) -> Result<()>

send a terminal to the specified soma