Struct organelle::Soma

pub struct Soma<M, R> where
    M: CellMessage,
    R: CellRole,  { /* fields omitted */ }

provides core convenience functions with little boilerplate

Methods

impl<M, R> Soma<M, R> where
M: CellMessage,
R: CellRole,

pub fn new(
inputs: Vec<Constraint<R>>,
outputs: Vec<Constraint<R>>
) -> Result<Self>

new soma with constraints and default user-defined state

pub fn update(&mut self, msg: Protocol<M, R>) -> Result<Option<Protocol<M, R>>>

update the soma's inputs and outputs, then verify constraints

if soma handles the given message, it consumes it, otherwise it is returned so that the cell can use it.

pub fn effector(&self) -> Result<&Effector<M, R>>

get the effector assigned to this cell

pub fn send_req_input(&self, dest: R, msg: M) -> Result<()>

convenience function for sending messages by role

pub fn send_req_output(&self, dest: R, msg: M) -> Result<()>

convenience function for sending messages by role

pub fn req_input(&self, role: R) -> Result<Handle>

get a RequireOne input

pub fn var_input(&self, role: R) -> Result<&Vec<Handle>>

get a Variadic input

pub fn req_output(&self, role: R) -> Result<Handle>

get a RequireOne output

pub fn var_output(&self, role: R) -> Result<&Vec<Handle>>

get a Variadic output