Trait organelle::Soma

pub trait Soma: Sized {
    type Signal: Signal;
    type Synapse: Synapse;
    fn update(self, _msg: Impulse<Self::Signal, Self::Synapse>) -> Result<Self> { ... }

    fn run(self) -> Result<()> { ... }
}

defines an interface for a soma of any type

generic across the user-defined message to be passed between somas and the user-defined roles for connections

Associated Types

type Signal: Signal

user-defined message to be passed between somas

type Synapse: Synapse

user-defined roles for connections

Provided Methods

fn update(self, _msg: Impulse<Self::Signal, Self::Synapse>) -> Result<Self>

apply any changes to the soma's state as a result of _msg

fn run(self) -> Result<()>

spin up an event loop and run soma