Trait organelle::Synapse

pub trait Synapse: Debug + Copy + Clone + Hash + PartialEq + Eq {
    type Terminal: Debug;
    type Dendrite: Debug;
    fn synapse(self) -> (Self::Terminal, Self::Dendrite);

    fn data() -> SynapseData { ... }
}

trait alias to express requirements of a Synapse type

Associated Types

type Terminal: Debug

terminals are the senders or outputs in a connection between somas

type Dendrite: Debug

dendrites are the receivers or inputs in a connection between somas

Required Methods

fn synapse(self) -> (Self::Terminal, Self::Dendrite)

form a synapse for this synapse into a terminal and dendrite

Provided Methods

fn data() -> SynapseData

get the data associated with the synapse