Struct organelle::Axon

pub struct Axon<S: Signal, Y: Synapse> { /* fields omitted */ }

provides core convenience functions with little boilerplate

Methods

impl<S: Signal, Y: Synapse> Axon<S, Y>

pub fn new(inputs: Vec<Dendrite<Y>>, outputs: Vec<Dendrite<Y>>) -> Result<Self>

new axon with dendrites and default user-defined state

pub fn update(&mut self, msg: Impulse<S, Y>) -> Result<Option<Impulse<S, Y>>>

update the axon's inputs and outputs, then verify dendrites

if axon handles the given message, it consumes it, otherwise it is returned so that the soma can use it.

pub fn effector(&self) -> Result<&Effector<S, Y>>

get the effector assigned to this soma

pub fn send_req_input(&self, dest: Y, msg: S) -> Result<()>

convenience function for sending messages by role

pub fn send_req_output(&self, dest: Y, msg: S) -> Result<()>

convenience function for sending messages by role

pub fn req_input(&self, role: Y) -> Result<Handle>

get a RequireOne input

pub fn var_input(&self, role: Y) -> Result<&Vec<Handle>>

get a Variadic input

pub fn req_output(&self, role: Y) -> Result<Handle>

get a RequireOne output

pub fn var_output(&self, role: Y) -> Result<&Vec<Handle>>

get a Variadic output